refactor: move modifyCow to its own file
